requirements.txt
numpy==1.24.3
tensorflow==2.13.0
opencv-python==4.11.0.86
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
Pillow==10.0.0
Flask==2.3.3
python-dotenv==1.0.0